Primeiro deploy da nova função de análise de curva
